Update pathfinder.bcgov
Trying to add wording around a common misconception that the *.pathfinder.bcgov route is secured.

While the pathfinder.bcgov VIP (142.34.143.180) may not be routable to an external IP, that is NOT a cluster route.  Both of the external VIPs are directing traffic to the SAME cluster ingress.  To properly protect a .bcgov route, it is neccessary to add IP Whitelists for the BCGov Subnets to the route object.